public api is strings
